updated ni_daq and launching for wavemeter daq
